refactor for blocks instructions
