+++ src/blocks_instructions/follow/follow_instructions.h
@@ -1,0 +1,5 @@
+void runDelay();
+void runRepeat();
+void runEndRepeat();
+void runBreak();
+void freeRepeatBegins();
