+++ src/blocks_instructions/functions/function_instructions.h
@@ -1,0 +1,4 @@
+
+
+void runGoToFunction();
+void runEndOfFunction();
