+++ src/blocks_instructions/logic/logic_instructions.cpp
@@ -1,0 +1,71 @@
+#include <stdint.h>
+#include "load_from_eeprom.h"
+#include "extern_libraries/linked_list/linked_list.h"
+#include "loginc_instructions.h"
+
+extern uint32_t currentOffset;
+
+int32_t getComparatorValue()
+{
+    int32_t firstValue = getInputValue(nextByte());
+    uint8_t operation = nextByte();
+    int32_t secondByte = getInputValue(nextByte());
+    switch (operation)
+    {
+    case 0:
+        return (firstValue == secondByte) ? 1 : 0;
+    case 1:
+        return (firstValue != secondByte) ? 1 : 0;
+    case 2:
+        return (firstValue < secondByte) ? 1 : 0;
+    case 3:
+        return (firstValue <= secondByte) ? 1 : 0;
+    case 4:
+        return (firstValue > secondByte) ? 1 : 0;
+    case 5:
+        return (firstValue >= secondByte) ? 1 : 0;
+    default:
+        return 0;
+    }
+}
+
+
+int32_t getLogicValue()
+{
+    uint8_t firstValue = getInputValue(nextByte()) == 0 ? 0 : 1;
+    uint8_t operation = nextByte();
+    uint8_t secondByte = getInputValue(nextByte()) == 0 ? 0 : 1;
+    switch (operation)
+    {
+    case 0:
+        return (firstValue && secondByte) ? 1 : 0;
+    case 1:
+        return (firstValue || secondByte) ? 1 : 0;
+    default:
+        return 0;
+    }
+}
+
+int32_t getNotValue()
+{
+    return (getInputValue(nextByte()) == 0) ? 1 : 0;
+}
+
+
+void runIf()
+{
+    int32_t conditionValue = getInputValue(nextByte());
+    uint8_t eosBytes[3];
+    for (uint8_t i = 0; i < 3; i++)
+    {
+        eosBytes[i] = nextByte();
+    }
+    uint32_t eos = (eosBytes[0] * 10000) + (eosBytes[1] * 100) + eosBytes[2];
+    if (conditionValue != 0)
+    {
+    }
+    else
+    {
+        currentOffset = eos;
+    }
+}
