+++ src/blocks_instructions/logic/loginc_instructions.h
@@ -1,0 +1,7 @@
+
+#include <stdint.h>
+
+int32_t getComparatorValue();
+int32_t getLogicValue();
+int32_t getNotValue();
+void runIf();
