+++ src/blocks_instructions/math/math_instructions.cpp
@@ -1,0 +1,47 @@
+#include <stdint.h>
+#include "load_from_eeprom.h"
+#include <Arduino.h>
+#include "math_instructions.h"
+
+extern uint32_t currentOffset;
+
+
+int32_t getMapValue()
+{
+    int32_t source = getInputValue(nextByte());
+    int32_t inMin = getInputValue(nextByte());
+    int32_t inMax = getInputValue(nextByte());
+    int32_t outMin = getInputValue(nextByte());
+    int32_t outMax = getInputValue(nextByte());
+    return map(source, inMin, inMax, outMin, outMax);
+}
+
+int32_t getRandomValue()
+{
+    int32_t from = getInputValue(nextByte());
+    int32_t to = getInputValue(nextByte());
+    return random(from, to);
+}
+
+
+int32_t getAritmeticValue()
+{
+    int32_t firstValue = getInputValue(nextByte());
+    uint8_t operation = nextByte();
+    int32_t secondByte = getInputValue(nextByte());
+    switch (operation)
+    {
+    case 0:
+        return firstValue + secondByte;
+    case 1:
+        return firstValue - secondByte;
+    case 2:
+        return firstValue * secondByte;
+    case 3:
+        return firstValue / secondByte;
+    case 4:
+        return firstValue % secondByte;
+    default:
+        return 0;
+    }
+}
