+++ src/blocks_instructions/math/math_instructions.h
@@ -1,0 +1,5 @@
+#include <stdint.h>
+
+int32_t getMapValue();
+int32_t getRandomValue();
+int32_t getAritmeticValue();
