+++ src/blocks_instructions/variables/variables_instructions.h
@@ -1,0 +1,7 @@
+#include <stdint.h>
+
+void variableEepromLoad();
+int32_t getVariableValue();
+void setVarValue(int32_t newValue, int32_t *value);
+void runSetVarValue(uint8_t id);
+void freeVariables();
