+++ src/value_conversion/value_conversion.h
@@ -1,0 +1,5 @@
+#include <stdint.h>
+
+uint8_t firstValue(uint32_t value);
+uint8_t secondValue(uint32_t value);
+uint8_t thirdValue(uint32_t value);
